+++ components/AssetExample.js
